+++ test/07compression.ts
@@ -15,9 +15,22 @@ describe("compression", () => {
       rfs => rfs.end("test\ntest\n")
     );
 
-    it("events", () => deq(events, { finish: 1, open: ["test.log/log", "test.log/log"], rotated: ["test.log/1"], rotation: 1, write: 1, ...v14() }));
+    it("events", () => deq(events, { finish: 1, open: ["test.log/log", "test.log/log"], rotated: ["test.log/1"], rotation: 1, stderr: [""], stdout: [""], write: 1, ...v14() }));
     it("file content", () => eq(readFileSync("test.log/log", "utf8"), ""));
     it("rotated file content", () => eq(gunzipSync(readFileSync("test.log/1")).toString(), "test\ntest\n"));
+  });
+
+  describe("external command error", () => {
+    const events = test(
+      {
+        filename: (time: number | Date, index: number) => (time ? `test.log/${index}` : "test.log/log"),
+        options:  { compress: (source, dest) => `echo ${source} ; >&2 echo ${dest} ; exit 23`, size: "10B" }
+      },
+      rfs => rfs.end("test\ntest\n")
+    );
+
+    it("events", () => deq(events, { error: [23], finish: 1, open: ["test.log/log"], rotation: 1, stderr: ["test.log/1\n"], stdout: ["test.log/log\n"], write: 1, ...v14() }));
+    it("file content", () => eq(readFileSync("test.log/log", "utf8"), "test\ntest\n"));
   });
 
   describe("custom external", () => {
@@ -29,7 +42,7 @@ describe("compression", () => {
       rfs => rfs.end("test\ntest\n")
     );
 
-    it("events", () => deq(events, { finish: 1, open: ["test.log", "test.log"], rotated: ["test1.log"], rotation: 1, write: 1, ...v14() }));
+    it("events", () => deq(events, { finish: 1, open: ["test.log", "test.log"], rotated: ["test1.log"], rotation: 1, stderr: [""], stdout: [""], write: 1, ...v14() }));
     it("file content", () => eq(readFileSync("test.log", "utf8"), ""));
     it("rotated file content", () => eq(gunzipSync(readFileSync("test1.log")).toString(), "test\ntest\n"));
   });
@@ -61,6 +74,6 @@ describe("compression", () => {
       rfs.write("test\ntest\n");
     });
 
-    it("events", () => deq(events, { close: 1, error: ["test"], finish: 1, open: ["test.log"], rotation: 1, write: 1 }));
+    it("events", () => deq(events, { close: 1, error: ["test"], finish: 1, open: ["test.log"], stderr: [undefined], stdout: [undefined], rotation: 1, write: 1 }));
   });
 });